apps/clawbot/requirements.txt
python-telegram-bot
python-dotenv
openai
google-genai
httpx
pyautogui
mss
Pillow
flask
